Fix accurate manifest durations after keyframe-aligned clipping

+++ vod/hls/hls_muxer.h
@@ -39,7 +39,7 @@ typedef struct {
 	// time offsets
 	uint64_t first_frame_time_offset;
 	uint64_t next_frame_time_offset;
-	int32_t clip_from_frame_offset;
+	int64_t clip_from_frame_offset;
 
 	// iframes simulation only
 	uint64_t segment_limit;
